Close sheet when page changes
